From Xavier Roche <roche@exalead.com> gsize is not always guinit32. Fixes
2003-04-09  Jody Goldberg <jody@gnome.org>

	From Xavier Roche <roche@exalead.com>
	* gsf/gsf-msole-utils.c (msole_prop_parse) : gsize is not always
	  guinit32.  Fixes sparc-64 problem.
	* gsf/gsf-infile-msvba.c (vba3_dir_read) : ditto.

2003-04-01  Jody Goldberg <jody@gnome.org>

	* gsf/gsf-impl-utils.h : Add utils for objects in plugins

+++ gsf/gsf-infile-msvba.c
@@ -239,7 +239,7 @@ vba3_dir_read (GsfInfileMSVBA *vba, GError **err)
 	};
 
 	GsfInput *dir;
-	size_t inflated_size;
+	int inflated_size;
 	guint8 *inflated;
 	unsigned offset, name_len, i, j;
 	char *name;
